refactor: Phase 3.2 — move 6 leaf routers to domain modules
Move woocommerce, amocrm, gsm, tts, stt, services routers from
app/routers/ to their domain modules. Replace db.integration imports
with direct domain service imports. Original files become thin facades.

Closes #509
